Etcetera

Documentation for Etcetera.

Etcetera.longest_repeated_substrings

Etcetera.longest_repeated_substrings — Method

longest_repeated_substrings(t,overlap_allowed)

tests the text t for the existence of longest repeated substrings (LRSs) of length 2 or more and returns:

the length of a longest repeated substring in t

the set of longest repeated substrings in t

per LRS a list of its occurences (starting positions).

If no such strings exists: an empty set is returned and length nothing.

Arguments

t: the text (String,Array,Tuple) to find longest repeated substrings in.

overlap_allowed: whether the occurences of a single LRS may overlap.

Examples

longest_repeated_substrings("ananas") == (3, Dict("ana" => [1, 3]))

longest_repeated_substrings("ananas"; overlap_allowed=false) == (2, Dict("an" => [1, 3], "na" => [2, 4]))